Resolve some compiler hints, add "affected row(s)" logic to the TZLoggingEvent (some drivers need to be revisited)
git-svn-id: http://svn.code.sf.net/p/zeoslib/code-0/branches/testing-7.3@6712 d681b350-854c-439c-a81d-c0ce20710203

+++ src/dbc/ZDbcASAResultSet.pas
@@ -1698,7 +1698,7 @@ end;
 function TZASAStream.Seek(const Offset: Int64; Origin: TSeekOrigin): Int64;
 begin
   if Origin = soEnd then
-    Result := PZASABlobStruct(FOwnerLob.FASASQLDA.sqlVar[FOwnerLob.FColumnIndex].sqlData).untrunc_len - OffSet
+    Result := Int64(PZASABlobStruct(FOwnerLob.FASASQLDA.sqlVar[FOwnerLob.FColumnIndex].sqlData).untrunc_len) - OffSet
   else if Origin = soCurrent then
     Result := FPosition + OffSet
   else
